chore(lap-427): change file and controller locations, add swagger decorator
