+++ apps/main-api/src/decorators/index.ts
@@ -1,2 +1,3 @@
 export * from "./current-user.decorator";
 export * from "./roles.decorator";
+export * from "./swagger.decorator";
