+++ apps/main-api/src/modules/item/inventory/index.ts
@@ -1,0 +1,2 @@
+export * from "./inventory.controller";
+export * from "./inventory.service";
