+++ apps/main-api/src/modules/item/item-effect/item-effect-factory.service.ts
@@ -2,13 +2,8 @@ import { IItem } from "@app/types/interfaces";
 import { DefaultItemEffect } from "./item-effect-abstract.service";
 import { RandomGiftItemEffect } from "./random-gift-item.service";
 import { LearnerProfile } from "@app/database";
+import { ItemName } from "@app/types/enums";
 
-enum ItemName {
-  STREAK_FREEZE = "STREAK_FREEZE",
-  RANDOM_GIFT = "RANDOM_GIFT",
-  ULTIMATE_TIME = "ULTIMATE_TIME",
-  IDENTIFICATION = "IDENTIFICATION",
-}
 export class ItemEffectFactoryService {
   createItemEffectService(item: IItem, learner: LearnerProfile) {
     switch (item.name) {
