+++ apps/main-api/src/modules/item/shop/index.ts
@@ -1,0 +1,2 @@
+export * from "./shop.controller";
+export * from "./shop.service";
