+++ apps/main-api/src/modules/notification/notification.controller.ts
@@ -1,27 +1,23 @@
-import { Body, Controller, HttpStatus, Post, UseGuards } from "@nestjs/common";
+import { Body, Controller, Post, UseGuards } from "@nestjs/common";
 import { FirebaseJwtAuthGuard } from "../../guards";
-import { ApiBearerAuth, ApiTags, ApiResponse } from "@nestjs/swagger";
+import { ApiBearerAuth, ApiTags, ApiOkResponse, ApiOperation } from "@nestjs/swagger";
 import { ICurrentUser } from "@app/types/interfaces";
-import { CurrentUser } from "../../decorators";
+import { ApiDefaultResponses, CurrentUser } from "../../decorators";
 import { NotificationService } from "./notification.service";
 import { FcmTokenDto } from "@app/types/dtos";
 
 @ApiTags("Notifications")
 @ApiBearerAuth()
-@ApiResponse({ status: 200, description: "Successfully" })
-@ApiResponse({ status: 400, description: "Bad request" })
-@ApiResponse({ status: 401, description: "Unauthorized" })
-@ApiResponse({ status: 403, description: "Forbidden" })
-@ApiResponse({ status: 500, description: "Internal server error" })
+@ApiDefaultResponses()
 @UseGuards(FirebaseJwtAuthGuard)
 @Controller("notifications")
 export class NotificationController {
   constructor(private readonly notificationService: NotificationService) {}
-  @Post("fcm-token")
-  @ApiResponse({
-    status: HttpStatus.OK,
+  @ApiOperation({ summary: "Upsert FCM token for the account" })
+  @ApiOkResponse({
     type: String,
   })
+  @Post("fcm-token")
   async upsertAccountToken(@CurrentUser() user: ICurrentUser, @Body() data: FcmTokenDto) {
     return this.notificationService.upsertAccountToken(user.userId, data.token);
   }
