+++ libs/types/src/enums/item-name.enum.ts
@@ -1,0 +1,6 @@
+export enum ItemName {
+  STREAK_FREEZE = "STREAK_FREEZE",
+  RANDOM_GIFT = "RANDOM_GIFT",
+  ULTIMATE_TIME = "ULTIMATE_TIME",
+  IDENTIFICATION = "IDENTIFICATION",
+}
