+++ libs/types/src/enums/random-gift-type.enum.ts
@@ -1,0 +1,4 @@
+export enum RandomGiftType {
+  CARROTS = "carrots",
+  ITEM = "item",
+}
