+++ libs/types/src/response-dtos/buckets/index.ts
@@ -1,0 +1,1 @@
+export * from "./get-bucket-response.dto";
