+++ libs/types/src/response-dtos/items/index.ts
@@ -1,0 +1,3 @@
+export * from "./get-items-response.dto";
+export * from "./profile-items-response.dto";
+export * from "./use-random-gift-response.dto";
