--- libs/types/src/response-dtos/success-response.dto.ts
@@ -1,15 +1,0 @@
-import { HttpStatus } from "@nestjs/common";
-import { ApiProperty } from "@nestjs/swagger";
-
-export class SuccessResponseDto<T> {
-  @ApiProperty({
-    example: HttpStatus.OK,
-  })
-  statusCode: number;
-
-  @ApiProperty()
-  message: string;
-
-  @ApiProperty()
-  data?: T;
-}
